dev_requirements.txt
setuptools >= 58.1.0
requests >= 2.28.0
wheel >= 0.37.1
coverage >= 6.4.1
twine >= 4.0.1
bumpversion >= 0.6.0 
tox >= 3.25.0
flake8 >= 4.0.1
pytest >= 7.1.2
pytest-cov >= 3.0.0
readme_renderer >= 35.0
docutils >= 0.13.1
azure-storage-blob >= 12.12.0
black >= 22.3.0
isort >= 5.10.1